requirements.txt
django
djangorestframework
powerline-status==2.6
